Added Win Sprites to Project, Added Tracks to Board, and Made Minor Edits
Added sprites for the win screen. Just shapes with crowns. Added the tracks to the board and moved the line renderers to line up with them. Just need to increment one of the y values and change the color of the line to the right material. Otherwise just edited the layout of stuff a little.